Feature/id how to use

+++ internal/domain/entity/authentication.go
@@ -1,11 +1,11 @@
 package entity
 
 type Authentication struct {
-	Id       int64
+	Id       int
 	Identify string
 	Token    string
 }
 
-func NewAuthentication(id int64, identify string, token string) *Authentication {
+func NewAuthentication(id int, identify string, token string) *Authentication {
 	return &Authentication{Id: id, Identify: identify, Token: token}
 }
