+++ internal/domain/entity/organization.go
@@ -1,11 +1,11 @@
 package entity
 
 type Organization struct {
-	ParentId int64
-	Id       int64
+	ParentId int
+	Id       int
 	Name     string
 }
 
-func NewOrganization(parentId, id int64, name string) *Organization {
+func NewOrganization(parentId, id int, name string) *Organization {
 	return &Organization{ParentId: parentId, Id: id, Name: name}
 }
