+++ internal/domain/repository/organization_members.go
@@ -7,9 +7,9 @@ import (
 )
 
 type OrganizationMembers interface {
-	Find(organizationId, individualId int64) (*entity.OrganizationMember, error)
-	Entry(organizationId, individualId int64, role value_object.Role) error
-	Leave(individualId int64, reason string) error
+	Find(organizationId, individualId int) (*entity.OrganizationMember, error)
+	Entry(organizationId, individualId int, role value_object.Role) error
+	Leave(individualId int, reason string) error
 	Members(organization *entity.Organization, individualRepository Individuals) (*[]entity.OrganizationMember, error)
 	IndividualsAssigned(individual *[]entity.Individual) (*[]entity.OrganizationMember, error)
 	Close()
