+++ internal/domain/usecase/authentication_test.go
@@ -21,11 +21,11 @@ func TestAuth(t *testing.T) {
 	authRep := mockRepository.NewMockAuthentications(ctrl)
 	organizationRep := mockRepository.NewMockOrganizations(ctrl)
 	memberRep := mockRepository.NewMockOrganizationMembers(ctrl2)
-	authRep.EXPECT().FindToken("TestUser").Return(int64(1), string(token), nil)
+	authRep.EXPECT().FindToken("TestUser").Return(1, string(token), nil)
 
 	individual := entity.NewIndividual(1, 0, 0, "TestUser")
 	memberRep.EXPECT().IndividualsAssigned(&[]entity.Individual{*individual}).Return(&[]entity.OrganizationMember{*entity.NewOrganizationMember(entity.NewOrganization(0, 1, "TestOrganization"), individual, value_object.ADMIN)}, nil)
-	organizationRep.EXPECT().Find(int64(1)).Return(entity.NewOrganization(int64(1), int64(1), "TestOrganization"), nil)
+	organizationRep.EXPECT().Find(1).Return(entity.NewOrganization(1, 1, "TestOrganization"), nil)
 
 	itr := NewAuthentication(authRep, organizationRep, memberRep)
 	id, _ := itr.Auth("TestOrganization", "TestUser", "password")
